AddConstToX86Reg:  pointer-to-int truncation fixed

+++ Source/RSP/Recompiler Ops.c
@@ -5447,7 +5447,7 @@ void Compile_Opcode_LRV ( void ) {
 	LeaSourceAndOffset(x86_EAX, x86_EAX, (size_t)&RSP_Vect[RSPOpC.rt].B[0]);
 	DecX86reg(x86_EAX);
 */
-	AddConstToX86Reg(x86_EAX, ((DWORD)&RSP_Vect[RSPOpC.rt].UB[0]) - 2);
+	AddConstToX86Reg(x86_EAX, ((size_t)&RSP_Vect[RSPOpC.rt].UB[0]) - 2);
 
 	CPU_Message("   Loop:");
 	Loop = RecompPos;
